Contrib script that generates a BIND c.f.ip6.arpa zone
If you ever wanted, for whatever reason, to generate a BIND reverse DNS lookup zone based on nodedb, then this script is somewhere to start. It isn't particularly robust, lies a bit by default in the NS/SOA records and can probably be easily defeated by unusual names or characters in the nodedb file, but maybe it will be useful to someone.